Fix bug using pd.asfreq()
Apparently, in the year 2011 there was not repetead values during the
same millisecond. However, years 2014 and 2019 had this characteristic.
Therefore I was having the wrong values and the frequency was not
taking in account the real trade signs.

I hade to use pd.resample to solve this problem. With this, all the
values are include in the computation.